Extract necessary function to utils
